fixup: leaks, docs were too much
