+++ openfeature/event_executor_test.go
@@ -12,9 +12,6 @@ import (
 	"go.uber.org/mock/gomock"
 )
 
-func init() {
-}
-
 // Requirement 5.1.1 The provider MAY define a mechanism for signaling the occurrence of one of a set of events,
 // including PROVIDER_READY, PROVIDER_ERROR, PROVIDER_CONFIGURATION_CHANGED and PROVIDER_STALE,
 // with a provider event details payload.
