+++ openfeature/openfeature_test.go
@@ -3,11 +3,9 @@ package openfeature
 import (
 	"errors"
 	"reflect"
-	"sync/atomic"
 	"testing"
 	"time"
 
-	"go.uber.org/goleak"
 	"go.uber.org/mock/gomock"
 )
 
@@ -841,151 +839,4 @@ func expectTimeout(t *testing.T, ch <-chan any, receiveMsg string) {
 	case <-ch:
 		t.Error(receiveMsg)
 	}
-}
-
-// TestNoGoroutineLeak verifies that the event executor's goroutine is properly
-// cleaned up after shutdown, preventing goroutine leaks.
-func TestNoGoroutineLeak(t *testing.T) {
-	// Start from a clean goroutine baseline and restore a working singleton
-	// afterwards (see startLeakTest).
-	startLeakTest(t)
-
-	// Verify no goroutines leak. The shutdown below is registered after this so
-	// it runs first (defers are LIFO), stopping the executor before we verify.
-	defer goleak.VerifyNone(t)
-	defer shutdownEventing()
-
-	// Create a new provider and trigger some events
-	eventingImpl := &ProviderEventing{
-		c: make(chan Event, 1),
-	}
-
-	eventingProvider := struct {
-		FeatureProvider
-		EventHandler
-	}{
-		NoopProvider{},
-		eventingImpl,
-	}
-
-	err := SetProvider(eventingProvider)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Add a handler to ensure the event system is active. The callback may run
-	// concurrently from multiple goroutines (executeHandler dispatches each
-	// handler in its own goroutine), so guard the flag against data races.
-	var callbackInvoked atomic.Bool
-	callback := func(details EventDetails) {
-		callbackInvoked.Store(true)
-	}
-	AddHandler(ProviderReady, &callback)
-
-	// Trigger an event
-	eventingImpl.Invoke(Event{
-		EventType: ProviderReady,
-		ProviderEventDetails: ProviderEventDetails{
-			Message: "Ready",
-		},
-	})
-
-	// Give the event system time to process
-	time.Sleep(100 * time.Millisecond)
-
-	// Verify the callback was invoked
-	if !callbackInvoked.Load() {
-		t.Error("callback should have been invoked")
-	}
-
-	// Perform shutdown - this cleans up goroutines but reinitializes the singleton
-	Shutdown()
-
-	// The deferred cleanup will shut down the reinitialized instance
-}
-
-// TestNoGoroutineLeakWithShutdownWithContext verifies that ShutdownWithContext
-// also properly cleans up goroutines.
-func TestNoGoroutineLeakWithShutdownWithContext(t *testing.T) {
-	// Start from a clean goroutine baseline and restore a working singleton
-	// afterwards (see startLeakTest).
-	startLeakTest(t)
-
-	// ShutdownWithContext reinitializes the global event executor via
-	// resetSingleton; shutdownEventing (registered after VerifyNone, so it runs
-	// first) stops that replacement before goleak verifies.
-	defer goleak.VerifyNone(t)
-	defer shutdownEventing()
-
-	eventingImpl := &ProviderEventing{c: make(chan Event, 1)}
-	err := SetProvider(struct {
-		FeatureProvider
-		EventHandler
-	}{NoopProvider{}, eventingImpl})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Use ShutdownWithContext instead of Shutdown
-	err = ShutdownWithContext(t.Context())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// goleak will verify no goroutines are leaked
-}
-
-// TestNoGoroutineLeakWithMultipleProviders verifies that goroutine cleanup
-// works correctly even when multiple providers are registered.
-func TestNoGoroutineLeakWithMultipleProviders(t *testing.T) {
-	// Start from a clean goroutine baseline and restore a working singleton
-	// afterwards (see startLeakTest).
-	startLeakTest(t)
-
-	// Verify no goroutines leak. The shutdown below is registered after this so
-	// it runs first (defers are LIFO), stopping the executor before we verify.
-	defer goleak.VerifyNone(t)
-	defer shutdownEventing()
-
-	// Set default provider
-	defaultProvider := &ProviderEventing{c: make(chan Event, 1)}
-	err := SetProvider(struct {
-		FeatureProvider
-		EventHandler
-	}{NoopProvider{}, defaultProvider})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Set named provider
-	namedProvider := &ProviderEventing{c: make(chan Event, 1)}
-	err = SetNamedProvider("test-domain", struct {
-		FeatureProvider
-		EventHandler
-	}{NoopProvider{}, namedProvider})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Trigger events
-	defaultProvider.Invoke(Event{
-		EventType: ProviderReady,
-		ProviderEventDetails: ProviderEventDetails{
-			Message: "Default ready",
-		},
-	})
-
-	namedProvider.Invoke(Event{
-		EventType: ProviderReady,
-		ProviderEventDetails: ProviderEventDetails{
-			Message: "Named ready",
-		},
-	})
-
-	time.Sleep(100 * time.Millisecond)
-
-	// Shutdown cleans up all goroutines and reinitializes
-	Shutdown()
-
-	// goleak will verify no goroutines are leaked
 }
